Move PopoverWrapper to the editor folder
